Updated project to use 26.1.2 and renamed references of v26_1_1 to v26_1 to make version bumps less jarring
