+++ build.gradle.kts
@@ -97,7 +97,7 @@ val projs = listOf(
     ":modern:v19_2",
     ":modern:v20_2",
     ":modern:v20_4",
-    ":deobsf:v26_1_1",
+    ":deobsf:v26_1",
 )
 
 val mergeMixins = tasks.register<MergeMixinConfigs>("mergeMixins") {
