+++ common/src/main/resources/META-INF/services/org.adde0109.pcf.PCFInitializer
@@ -20,16 +20,16 @@ org.adde0109.pcf.v20_4.forge.AdapterRegistryInit
 org.adde0109.pcf.v20_4.forge.CrossStitchInit
 org.adde0109.pcf.v20_4.forge.ReloadConfigInit
 
-org.adde0109.pcf.v26_1_1.forge.DisplayTestInit
-org.adde0109.pcf.v26_1_1.forge.DisplayTestReg
-org.adde0109.pcf.v26_1_1.forge.RegisterConfig
-org.adde0109.pcf.v26_1_1.forge.ReloadConfigInit
+org.adde0109.pcf.v26_1.forge.DisplayTestInit
+org.adde0109.pcf.v26_1.forge.DisplayTestReg
+org.adde0109.pcf.v26_1.forge.RegisterConfig
+org.adde0109.pcf.v26_1.forge.ReloadConfigInit
 
 org.adde0109.pcf.v20_2.neoforge.CrossStitchInit
 org.adde0109.pcf.v20_2.neoforge.RegisterConfig
 
-org.adde0109.pcf.v26_1_1.neoforge.RegisterConfig
-org.adde0109.pcf.v26_1_1.neoforge.ReloadConfigInit
+org.adde0109.pcf.v26_1.neoforge.RegisterConfig
+org.adde0109.pcf.v26_1.neoforge.ReloadConfigInit
 
-org.adde0109.pcf.v26_1_1.AdapterRegistryInit
-org.adde0109.pcf.v26_1_1.CrossStitchInit
+org.adde0109.pcf.v26_1.AdapterRegistryInit
+org.adde0109.pcf.v26_1.CrossStitchInit
