+++ deobsf/v26_1/gradle.properties
@@ -1,0 +1,3 @@
+minecraft_version=26.1.2
+forge_version=64.0.0
+neoforge_version=0-beta
