+++ deobsf/v26_1/src/main/java/org/adde0109/pcf/v26_1/forwarding/network/QueryAnswerPayloadSearge.java
@@ -1,0 +1,15 @@
+package org.adde0109.pcf.v26_1.forwarding.network;
+
+import net.minecraft.network.FriendlyByteBuf;
+
+import org.jspecify.annotations.NonNull;
+
+/** Dummy interface so QueryAnswerPayload can work in Searge-mapped environments. */
+public sealed interface QueryAnswerPayloadSearge permits MCQueryAnswerPayload {
+    void write(final @NonNull FriendlyByteBuf buf);
+
+    @SuppressWarnings("unused")
+    default void m_295630_(final @NonNull FriendlyByteBuf buf) {
+        this.write(buf);
+    }
+}
