--- deobsf/v26_1_1/gradle.properties
@@ -1,3 +1,0 @@
-minecraft_version=26.1.1
-forge_version=63.0.1
-neoforge_version=14-beta
