--- deobsf/v26_1_1/src/main/java/org/adde0109/pcf/mixin/v26_1_1/forwarding/modern/ServerLoginPacketListenerImplQueryMixin.java
@@ -1,30 +1,0 @@
-package org.adde0109.pcf.mixin.v26_1_1.forwarding.modern;
-
-import static org.adde0109.pcf.forwarding.modern.ModernForwarding.handleCustomQueryPacket;
-
-import dev.neuralnexus.taterapi.meta.Mappings;
-import dev.neuralnexus.taterapi.meta.anno.AConstraint;
-import dev.neuralnexus.taterapi.meta.anno.Versions;
-import dev.neuralnexus.taterapi.meta.enums.MinecraftVersion;
-
-import net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket;
-import net.minecraft.server.network.ServerLoginPacketListenerImpl;
-
-import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
-import org.jspecify.annotations.NonNull;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@AConstraint(mappings = Mappings.MOJANG, version = @Versions(min = MinecraftVersion.V20_2))
-@Mixin(ServerLoginPacketListenerImpl.class)
-public abstract class ServerLoginPacketListenerImplQueryMixin
-        implements ServerLoginPacketListenerBridge {
-    @Inject(method = "handleCustomQueryPacket", at = @At("HEAD"), cancellable = true)
-    private void onHandleCustomQueryPacket(
-            final @NonNull ServerboundCustomQueryAnswerPacket packet,
-            final @NonNull CallbackInfo ci) {
-        handleCustomQueryPacket(this, packet.transactionId(), packet, ci);
-    }
-}
