--- deobsf/v26_1_1/src/main/java/org/adde0109/pcf/v26_1_1/forwarding/network/CCustomQueryPacketAdapter.java
@@ -1,52 +1,0 @@
-package org.adde0109.pcf.v26_1_1.forwarding.network;
-
-import static dev.neuralnexus.taterapi.resources.Identifier.identifier;
-
-import dev.neuralnexus.taterapi.meta.Constraint;
-import dev.neuralnexus.taterapi.meta.MinecraftVersions;
-import dev.neuralnexus.taterapi.network.protocol.login.ClientboundCustomQueryPacket;
-import dev.neuralnexus.taterapi.network.protocol.login.custom.CustomQueryPayload;
-import dev.neuralnexus.taterapi.serialization.Result;
-import dev.neuralnexus.taterapi.serialization.codecs.ReversibleCodec;
-
-import io.netty.buffer.Unpooled;
-
-import net.minecraft.network.FriendlyByteBuf;
-
-import org.adde0109.pcf.v20_2.neoforge.forwarding.network.MCQueryPayload_RL;
-
-public final class CCustomQueryPacketAdapter
-        implements ReversibleCodec<
-                net.minecraft.network.protocol.login.ClientboundCustomQueryPacket,
-                ClientboundCustomQueryPacket> {
-    public static final CCustomQueryPacketAdapter INSTANCE = new CCustomQueryPacketAdapter();
-
-    @Override
-    public Result<ClientboundCustomQueryPacket> encode(
-            final net.minecraft.network.protocol.login.ClientboundCustomQueryPacket object) {
-        final FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-        object.payload().write(buf);
-        return Result.success(
-                new ClientboundCustomQueryPacket(
-                        object.transactionId(),
-                        CustomQueryPayload.codec(object.payload().id().toString())
-                                .decode(buf.slice())));
-    }
-
-    @Override
-    public Result<net.minecraft.network.protocol.login.ClientboundCustomQueryPacket> decode(
-            final ClientboundCustomQueryPacket object) {
-        final net.minecraft.network.protocol.login.custom.CustomQueryPayload payload;
-        if (Constraint.lessThan(MinecraftVersions.V21_11).result()) {
-            payload =
-                    new MCQueryPayload_RL(
-                            identifier(object.payload().id()), object.payload().data());
-        } else {
-            payload =
-                    new MCQueryPayload(identifier(object.payload().id()), object.payload().data());
-        }
-        return Result.success(
-                new net.minecraft.network.protocol.login.ClientboundCustomQueryPacket(
-                        object.transactionId(), payload));
-    }
-}
