--- deobsf/v26_1_1/src/main/java/org/adde0109/pcf/v26_1_1/forwarding/network/MCQueryAnswerPayload.java
@@ -1,16 +1,0 @@
-package org.adde0109.pcf.v26_1_1.forwarding.network;
-
-import io.netty.buffer.ByteBuf;
-
-import net.minecraft.network.FriendlyByteBuf;
-
-import org.jspecify.annotations.NonNull;
-
-public record MCQueryAnswerPayload(@NonNull ByteBuf data)
-        implements net.minecraft.network.protocol.login.custom.CustomQueryAnswerPayload,
-                QueryAnswerPayloadSearge {
-    @Override
-    public void write(final @NonNull FriendlyByteBuf buf) {
-        buf.writeBytes(this.data.slice());
-    }
-}
