--- deobsf/v26_1_1/src/main/java/org/adde0109/pcf/v26_1_1/forwarding/network/SCustomQueryAnswerPacketAdapter.java
@@ -1,46 +1,0 @@
-package org.adde0109.pcf.v26_1_1.forwarding.network;
-
-import dev.neuralnexus.taterapi.network.protocol.login.ServerboundCustomQueryAnswerPacket;
-import dev.neuralnexus.taterapi.network.protocol.login.custom.CustomQueryAnswerPayload;
-import dev.neuralnexus.taterapi.serialization.Result;
-import dev.neuralnexus.taterapi.serialization.codecs.ReversibleCodec;
-
-import io.netty.buffer.Unpooled;
-
-import net.minecraft.network.FriendlyByteBuf;
-
-public final class SCustomQueryAnswerPacketAdapter
-        implements ReversibleCodec<
-                net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket,
-                ServerboundCustomQueryAnswerPacket> {
-    public static final SCustomQueryAnswerPacketAdapter INSTANCE =
-            new SCustomQueryAnswerPacketAdapter();
-
-    @Override
-    public Result<ServerboundCustomQueryAnswerPacket> encode(
-            final net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket object) {
-        final int transactionId = object.transactionId();
-        if (object.payload() == null) {
-            return Result.success(new ServerboundCustomQueryAnswerPacket(transactionId));
-        }
-        final FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-        object.payload().write(buf);
-        return Result.success(
-                new ServerboundCustomQueryAnswerPacket(
-                        transactionId,
-                        CustomQueryAnswerPayload.codec(transactionId).decode(buf.slice())));
-    }
-
-    @Override
-    public Result<net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket> decode(
-            final ServerboundCustomQueryAnswerPacket object) {
-        if (object.payload() == null) {
-            return Result.success(
-                    new net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket(
-                            object.transactionId(), null));
-        }
-        return Result.success(
-                new net.minecraft.network.protocol.login.ServerboundCustomQueryAnswerPacket(
-                        object.transactionId(), new MCQueryAnswerPayload(object.payload().data())));
-    }
-}
