+++ modern/v20_4/build.gradle.kts
@@ -21,7 +21,7 @@ unimined.minecraft(forge) {
 dependencies {
     forgeCompileOnly(srcSetAsDep(":modern:v16_5", "forge"))
     forgeCompileOnly(srcSetAsDep(":modern:v17_1", "forge"))
-    forgeCompileOnly(project(":deobsf:v26_1_1"))
+    forgeCompileOnly(project(":deobsf:v26_1"))
     forgeCompileOnly(project(":common"))
 }
